+++ apps/api/src/app/portfolio/current-rate.service.spec.ts
@@ -73,7 +73,7 @@ describe('CurrentRateService', () => {
 
   beforeAll(async () => {
     dataProviderService = new DataProviderService(null, [], null);
-    exchangeRateDataService = new ExchangeRateDataService(null, null);
+    exchangeRateDataService = new ExchangeRateDataService(null, null, null);
     marketDataService = new MarketDataService(null);
 
     await exchangeRateDataService.initialize();